Auto-Melee fix
The constant predefined MASK_SHOT was broken in the latest beta update of lmaobox.net. The library LnxLib was using that pre defined constant for doing visible only checks, and since MASK_SHOT was responding with a nil value, it would not work.
The issue has been resolved on the luas end.